test: AdminViewController 테스트 케이스 작성
